backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
sqlmodel>=0.0.16
alembic>=1.13.1
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
openai>=1.12.0
